(#157) changed calendar route to index and added util
